+++ src/main/java/com/github/leopoko/tacz_attributes/mixin/EntityKineticBulletMixin.java
@@ -2,10 +2,8 @@ package com.github.leopoko.tacz_attributes.mixin;
 
 import com.github.leopoko.tacz_attributes.attribute.CustomAttributes;
 import com.github.leopoko.tacz_attributes.attribute.GunType;
-import com.github.leopoko.tacz_attributes.util.FireModeHelper;
+import com.github.leopoko.tacz_attributes.util.DamageModifierHelper;
 import com.github.leopoko.tacz_attributes.util.GunTypeResolver;
-import com.tacz.guns.api.entity.IGunOperator;
-import com.tacz.guns.api.item.gun.FireMode;
 import com.tacz.guns.entity.EntityKineticBullet;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.Entity;
@@ -60,39 +58,7 @@ public abstract class EntityKineticBulletMixin extends Projectile {
         Entity owner = this.getOwner();
         if (!(owner instanceof LivingEntity shooter)) return damage;
 
-        GunType gunType = GunTypeResolver.resolve(this.gunId);
-
-        // 全体ダメージ倍率
-        double globalModifier = tacz_attributes$getAttributeValue(shooter, CustomAttributes.GUN_DAMAGE.get());
-
-        // 銃種別ダメージ倍率
-        double typeModifier = 1.0;
-        if (gunType != null) {
-            typeModifier = tacz_attributes$getAttributeValue(shooter, gunType.getDamageAttribute().get());
-        }
-
-        // ADS / 腰撃ちダメージ倍率
-        boolean isAds = IGunOperator.fromLivingEntity(shooter).getSynIsAiming();
-        double adsHipGlobal;
-        double adsHipType = 1.0;
-        if (isAds) {
-            adsHipGlobal = tacz_attributes$getAttributeValue(shooter, CustomAttributes.ADS_DAMAGE.get());
-            if (gunType != null) {
-                adsHipType = tacz_attributes$getAttributeValue(shooter, gunType.getAdsDamageAttribute().get());
-            }
-        } else {
-            adsHipGlobal = tacz_attributes$getAttributeValue(shooter, CustomAttributes.HIP_FIRE_DAMAGE.get());
-            if (gunType != null) {
-                adsHipType = tacz_attributes$getAttributeValue(shooter, gunType.getHipFireDamageAttribute().get());
-            }
-        }
-
-        // 射撃モード別ダメージ倍率
-        FireMode fireMode = FireModeHelper.getFireMode(shooter.getMainHandItem());
-        double fireModeGlobal = FireModeHelper.getAttributeValue(shooter, FireModeHelper.getGlobalDamageAttribute(fireMode));
-        double fireModeType = FireModeHelper.getAttributeValue(shooter, FireModeHelper.getTypeDamageAttribute(gunType, fireMode));
-
-        double combinedModifier = globalModifier * typeModifier * adsHipGlobal * adsHipType * fireModeGlobal * fireModeType;
+        double combinedModifier = DamageModifierHelper.calculateDamageModifier(shooter, this.gunId);
         if (combinedModifier == 1.0) return damage;
 
         return (float) (damage * combinedModifier);